# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Groundwork Engineering Evidence API",
    "version": "3.0.0",
    "description": "Workspace-scoped evidence, change analysis, connector, policy, release, document, and grounded retrieval API."
  },
  "security": [{ "bearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "parameters": {
      "WorkspaceId": { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
      "ConnectionId": { "name": "connectionId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
      "ChangeId": { "name": "changeId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
      "ReleaseId": { "name": "releaseId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
    },
    "schemas": {
...
      "EvidenceItem": {
        "type": "object",
        "required": ["externalId", "artifactType", "title", "sourceVersion", "content"],
        "properties": {
          "externalId": { "type": "string" },
          "artifactType": { "type": "string" },
          "title": { "type": "string" },
          "canonicalUrl": { "type": "string", "format": "uri" },
          "sourceVersion": { "type": "string" },
          "content": { "type": "string" },
          "metadata": { "type": "object", "additionalProperties": true },
          "sourceAcl": { "type": "object", "additionalProperties": true }
        }
      },
...
      "PolicyRequest": {
        "type": "object",
        "required": ["name", "ruleType", "severity", "definition"],
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "ruleType": { "type": "string" },
          "severity": { "type": "string", "enum": ["INFO", "LOW", "MEDIUM", "HIGH", "CRITICAL"] },
          "definition": { "type": "object", "required": ["findingCategory"], "properties": { "findingCategory": { "type": "string" } } },
          "enabled": { "type": "boolean" }
        }
      },
      "ReleaseRequest": {
        "type": "object",
        "required": ["name", "repositoryFullName", "headRef", "changeSetIds"],
        "properties": {
          "name": { "type": "string" },
          "repositoryFullName": { "type": "string", "pattern": "^[^/]+/[^/]+$" },
          "baseRef": { "type": "string" },
          "headRef": { "type": "string" },
          "changeSetIds": { "type": "array", "minItems": 1, "items": { "type": "string", "format": "uuid" } }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/workspaces/{workspaceId}/connections/{connectionId}/evidence": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }, { "$ref": "#/components/parameters/ConnectionId" }],
      "post": {
        "summary": "Ingest evidence through the provider-neutral boundary",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["artifacts"], "properties": { "artifacts": { "type": "array", "items": { "$ref": "#/components/schemas/EvidenceItem" } }, "reconcile": { "type": "boolean" } } } } } },
        "responses": { "200": { "description": "Sync result" } }
      }
    },
...
    "/api/workspaces/{workspaceId}/evidence": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }],
      "get": { "summary": "List evidence artifacts", "parameters": [{ "name": "type", "in": "query", "schema": { "type": "string" } }, { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 200 } }], "responses": { "200": { "description": "Artifacts" } } }
    },
    "/api/workspaces/{workspaceId}/evidence/search": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }],
      "get": { "summary": "Hybrid search with optional graph expansion", "parameters": [{ "name": "query", "in": "query", "required": true, "schema": { "type": "string" } }, { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 50 } }, { "name": "expandGraph", "in": "query", "schema": { "type": "boolean", "default": true } }], "responses": { "200": { "description": "Cited evidence hits" } } }
    },
    "/api/workspaces/{workspaceId}/evidence/{artifactId}": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }, { "name": "artifactId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
      "get": { "summary": "Get artifact versions and relationships", "responses": { "200": { "description": "Artifact detail" } } }
    },
...
    "/api/workspaces/{workspaceId}/policies": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }],
      "get": { "summary": "List versioned evidence policies", "responses": { "200": { "description": "Policies" } } },
      "post": {
        "summary": "Create a policy version",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PolicyRequest" } } } },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/api/workspaces/{workspaceId}/policies/{policyId}/activation": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }, { "name": "policyId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
      "patch": { "summary": "Activate/deactivate a specific policy version", "responses": { "200": { "description": "Activation changed" } } }
    },
    "/api/workspaces/{workspaceId}/changes/{changeId}/policy-exceptions": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }, { "$ref": "#/components/parameters/ChangeId" }],
      "post": { "summary": "Create a reasoned expiring policy exception", "responses": { "201": { "description": "Exception created" } } }
    },
    "/api/workspaces/{workspaceId}/releases": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }],
      "get": { "summary": "List release evidence records", "responses": { "200": { "description": "Releases" } } },
      "post": {
        "summary": "Freeze selected changes into a release record",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReleaseRequest" } } } },
        "responses": { "200": { "description": "Release record" } }
      }
    },
    "/api/workspaces/{workspaceId}/releases/{releaseId}": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }, { "$ref": "#/components/parameters/ReleaseId" }],
      "get": { "summary": "Get release record and live digest verification", "responses": { "200": { "description": "Release detail" } } }
    },
    "/api/workspaces/{workspaceId}/releases/{releaseId}/export": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }, { "$ref": "#/components/parameters/ReleaseId" }],
      "get": { "summary": "Download canonical JSON release evidence", "responses": { "200": { "description": "JSON export" } } }
    },
    "/api/workspaces/{workspaceId}/releases/{releaseId}/export.html": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }, { "$ref": "#/components/parameters/ReleaseId" }],
      "get": { "summary": "Download HTML release evidence", "responses": { "200": { "description": "HTML export" } } }
    },
    "/api/workspaces/{workspaceId}/releases/{releaseId}/export.pdf": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }, { "$ref": "#/components/parameters/ReleaseId" }],
      "get": { "summary": "Download PDF release evidence", "responses": { "200": { "description": "PDF export" } } }
    },
...
    "/api/workspaces/{workspaceId}/demo/evidence": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceId" }],
      "post": { "summary": "Seed a realistic cross-source evidence demo", "responses": { "201": { "description": "Seeded" } } }
    },
...
  }
}